refactor(trusty-common): split retrieval/mod.rs into submodules under 500-SLOC cap (#607)
Split the 830-SLOC retrieval/mod.rs into 5 focused sibling files:

- embedder.rs (33 SLOC): SHARED_EMBEDDER, shared_embedder(), seed_shared_embedder_with_mock()
- types.rs (68 SLOC): L0Identity, L1Essential, RecallResult, L1_CAP, RememberOptions, CrossPalaceResult, RetrievalLayers
- handle.rs (400 SLOC): PalaceHandle struct + new, open, flush, remember, remember_with_options, forget, list_drawers, purge_expired, rebuild_closets, refresh_l1, log_recall
- layers.rs (347 SLOC): retrieve_l0_l1, rescore_l1_by_similarity, retrieve_l2, retrieve_l3, expand_query, recall, recall_deep, recall_with_default_embedder, recall_deep_with_default_embedder, recall_across_palaces, recall_across_palaces_with_default_embedder, room_to_uuid, uuid_prefix_eq, dedup_extend
- mod.rs (20 SLOC): thin re-export facade

Zero behavior change; zero public API change.
log_recall moved to PalaceHandle method; room_to_uuid made pub for
handle.rs and layers.rs access from sibling submodules.
